refactor: move files from utils to internal
