+++ frontend/Makefile
@@ -5,10 +5,9 @@ IMAGE_NAME ?= boshi-frontend
 TAG ?= latest
 BUILD_NAME ?= $(DOCKER_REGISTRY)/$(IMAGE_NAME)
 PLATFORM ?= linux/arm64
+MAIN_TARGET = lib/main.dart
 
 DEVICE_ID ?= chrome
-MAIN_STAGING = lib/main_staging.dart
-DEV_STAGING = lib/main_development.dart
 FRONTEND_PORT=3000
 
 ifneq (,$(wildcard ./.env))
@@ -18,10 +17,7 @@ endif
 
 run:
 	scripts/start.sh &
-	flutter run -d ${DEVICE_ID} --target ${DEV_STAGING} --pid-file /tmp/flutter.pid --web-hostname 127.0.0.1 --web-port ${FRONTEND_PORT}
-
-staging:
-	flutter run -d ${DEVICE_ID} --target ${MAIN_STAGING}
+	flutter run -d $(DEVICE_ID) --target $(MAIN_TARGET) --pid-file /tmp/flutter.pid --web-hostname 127.0.0.1 --web-port $(FRONTEND_PORT)
 
 types:
 	dart run build_runner watch -d
@@ -31,7 +27,7 @@ docker:
 	docker buildx build \
 	--build-arg NGINX_HOST=$(HOST) \
 	--build-arg BACKEND_BASE_URL=$(BACKEND_BASE_URL) \
-	--build-arg BUILD_TARGET=$(MAIN_STAGING) \
+	--build-arg BUILD_TARGET=$(MAIN_TARGET) \
 	--build-arg CLIENT_METADATA_PATH=$(CLIENT_METADATA_PATH) \
 	--build-arg FEED_GENERATOR_URI=$(FEED_GENERATOR_URI) \
 	--build-arg PROD=$(PROD) \
