+++ frontend/lib/data/services/api/api_client.dart
@@ -4,10 +4,10 @@ import 'package:atproto/atproto.dart';
 import 'package:atproto/atproto_oauth.dart';
 import 'package:atproto/core.dart';
 import 'package:bluesky/bluesky.dart' as bsky;
-import 'package:frontend/config/environment.dart';
-import 'package:frontend/shared/exceptions/missing_env.dart';
+import 'package:frontend/internal/config/environment.dart';
+import 'package:frontend/internal/exceptions/missing_env.dart';
 import 'package:frontend/shared/models/reaction_payload/reaction_payload.dart';
-import 'package:frontend/utils/result.dart';
+import 'package:frontend/internal/result/result.dart';
 import 'package:frontend/shared/models/report/report.dart' as boshi_report;
 import 'package:frontend/domain/models/post/post.dart';
 import 'package:frontend/domain/models/user/user.dart';
@@ -17,14 +17,14 @@ import 'package:frontend/shared/models/post/post.dart' as post_request;
 import 'package:http/http.dart' as http;
 import 'package:frontend/data/models/requests/add_email/add_email.dart';
 import 'package:frontend/data/models/requests/verify_code/verify_code.dart';
-import 'package:frontend/shared/exceptions/verification_code_already_set_exception.dart';
-import 'package:frontend/shared/exceptions/code_not_found_exception.dart';
-import 'package:frontend/shared/exceptions/already_verified_exception.dart';
-import 'package:frontend/shared/exceptions/user_not_found_exception.dart';
+import 'package:frontend/internal/exceptions/verification_code_already_set_exception.dart';
+import 'package:frontend/internal/exceptions/code_not_found_exception.dart';
+import 'package:frontend/internal/exceptions/already_verified_exception.dart';
+import 'package:frontend/internal/exceptions/user_not_found_exception.dart';
 import 'package:frontend/data/models/responses/verification_status/verification_status.dart';
 import 'package:frontend/data/models/responses/verification_code_ttl/verification_code_ttl.dart';
 import 'package:frontend/shared/models/mock_data/feed/feed.dart';
-import 'package:frontend/utils/logger.dart';
+import 'package:frontend/internal/logger/logger.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 Future<void> _setSessionVars(
