+++ frontend/lib/internal/config/dependencies.dart
