+++ frontend/lib/internal/exceptions/already_verified_exception.dart
