+++ frontend/lib/internal/exceptions/oauth_unauthorized_exception.dart
