+++ frontend/lib/internal/exceptions/user_not_found_exception.dart
