+++ frontend/lib/internal/exceptions/verification_code_already_set_exception.dart
