+++ frontend/lib/internal/result/result.dart
