+++ frontend/lib/main.dart
@@ -1,29 +1,29 @@
 import 'package:flutter/material.dart';
-import 'package:frontend/ui/post/view_model/post_viewmodel.dart';
-import 'package:frontend/ui/post/widgets/post_screen.dart';
+import 'ui/post/view_model/post_viewmodel.dart';
+import 'ui/post/widgets/post_screen.dart';
 
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 import 'package:shadcn_ui/shadcn_ui.dart';
 
-import 'package:frontend/ui/login/widgets/login_screen.dart';
-import 'package:frontend/ui/login/widgets/oauth_callback_screen.dart';
-import 'package:frontend/ui/login/view_model/login_viewmodel.dart';
+import 'ui/login/widgets/login_screen.dart';
+import 'ui/login/widgets/oauth_callback_screen.dart';
+import 'ui/login/view_model/login_viewmodel.dart';
 
-import 'package:frontend/ui/home/view_model/home_viewmodel.dart';
-import 'package:frontend/ui/home/widgets/home_screen.dart';
-import 'package:frontend/ui/signup/widgets/email_register_screen.dart';
-import 'package:frontend/ui/signup/view_model/email_register_viewmodel.dart';
-import 'package:frontend/ui/signup/widgets/email_verification_screen.dart';
-import 'package:frontend/ui/signup/view_model/email_verification_viewmodel.dart';
+import 'ui/home/view_model/home_viewmodel.dart';
+import 'ui/home/widgets/home_screen.dart';
+import 'ui/signup/widgets/email_register_screen.dart';
+import 'ui/signup/view_model/email_register_viewmodel.dart';
+import 'ui/signup/widgets/email_verification_screen.dart';
+import 'ui/signup/view_model/email_verification_viewmodel.dart';
 
 import 'package:frontend/data/repositories/atproto/atproto_repository.dart';
 
-import 'package:frontend/utils/auth.dart';
-import 'package:frontend/config/environment.dart';
 import 'package:logger/logger.dart';
 import 'package:flutter_web_plugins/url_strategy.dart';
-import 'config/dependencies.dart';
+import 'internal/guards/guards.dart';
+import 'internal/config/environment.dart';
+import 'internal/config/dependencies.dart';
 
 void main() {
   Logger.level = Level.all;
