+++ frontend/lib/ui/login/view_model/login_viewmodel.dart
@@ -2,9 +2,9 @@ import 'package:frontend/data/repositories/atproto/atproto_repository.dart';
 
 import 'package:flutter/widgets.dart';
 
-import 'package:frontend/utils/result.dart';
-import 'package:frontend/utils/command.dart';
-import 'package:frontend/utils/logger.dart';
+import 'package:frontend/internal/result/result.dart';
+import 'package:frontend/internal/command/command.dart';
+import 'package:frontend/internal/logger/logger.dart';
 
 import 'package:frontend/ui/models/login/login.dart';
