+++ frontend/lib/ui/signup/view_model/email_register_viewmodel.dart
@@ -1,7 +1,7 @@
 import 'package:frontend/data/repositories/atproto/atproto_repository.dart';
 import 'package:flutter/material.dart';
-import 'package:frontend/utils/result.dart';
-import 'package:frontend/utils/command.dart';
+import 'package:frontend/internal/result/result.dart';
+import 'package:frontend/internal/command/command.dart';
 
 class EmailRegisterViewModel extends ChangeNotifier {
   EmailRegisterViewModel({
